WIP: Still need to break out header and setup shared styles. Then realign and refactor tests
