+++ services/ui-src/src/routes/index.tsx
@@ -7,6 +7,7 @@ export { NotFound } from "./NotFound/NotFound";
 export { Profile } from "./Profile/Profile";
 // report
 export { GetStarted } from "./GetStarted/GetStarted";
-export { Dashboard } from "./Dashboard/Dashboard";
-export { DashboardList } from "./Dashboard/DashboardList"
+export { DashboardPage } from "./Dashboard/DashboardPage";
+export { DashboardList } from "./Dashboard/DashboardList";
+export { MobileDashboardList } from "./Dashboard/MobileDashboardList";
 export { ReviewSubmit } from "./ReviewSubmit/ReviewSubmit";
